fix: only add PSC ipType if PSC is enabled
