+++ src/sqladmin-fetcher.ts
@@ -146,10 +146,31 @@ export class SQLAdminFetcher {
       });
     }
 
-    const ipAddresses = parseIpAddresses(
-      res.data.ipAddresses,
-      res.data.dnsName
-    );
+    const ipAddresses: IpAddresses = {};
+    if (res.data.ipAddresses) {
+      for (const ip of res.data.ipAddresses) {
+        if (ip.type === 'PRIMARY' && ip.ipAddress) {
+          ipAddresses.public = ip.ipAddress;
+        }
+        if (ip.type === 'PRIVATE' && ip.ipAddress) {
+          ipAddresses.private = ip.ipAddress;
+        }
+      }
+    }
+
+    // Resolve dnsName into IP address for PSC enabled instances.
+    // Note that we have to check for PSC enablement also because CAS instances
+    // also set the dnsName field.
+    if (res.data.dnsName && res.data.pscEnabled) {
+      ipAddresses.psc = res.data.dnsName;
+    }
+
+    if (!ipAddresses.public && !ipAddresses.private && !ipAddresses.psc) {
+      throw new CloudSQLConnectorError({
+        message: 'Cannot connect to instance, it has no supported IP addresses',
+        code: 'ENOSQLADMINIPADDRESS',
+      });
+    }
 
     const {serverCaCert} = res.data;
     if (!serverCaCert || !serverCaCert.cert || !serverCaCert.expirationTime) {
